+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./ids.js";
 export * from "./hostnames.js";
+export * from "./site-routing.js";
 export * from "./crypto.js";
 export * from "./functions.js";
 export * from "./function-templates.js";